Add files via upload
The primary goal of this project is to design and implement an E-commerce Product Page. This page will showcase a single product, emphasizing the importance of clean and effective web design. The use of HTML and CSS will be essential in structuring the content and styling the page to create an aesthetically pleasing and user-friendly experience.